refactor: move IUnSub interface to a dedicated unsub package
